+++ src/quacc/calculators/vasp/vasp.py
@@ -2,9 +2,8 @@
 
 from __future__ import annotations
 
-import inspect
 import os
-import sys
+import subprocess
 from pathlib import Path
 from typing import TYPE_CHECKING
 
@@ -13,20 +12,24 @@ from ase.calculators.vasp import Vasp as Vasp_
 from ase.calculators.vasp import setups as ase_setups
 from ase.constraints import FixAtoms
 
-from quacc.calculators.vasp import vasp_custodian
 from quacc.calculators.vasp.io import load_vasp_yaml_calc
 from quacc.calculators.vasp.params import (
     get_param_swaps,
+    get_pmg_input_set_params,
+    normalize_params,
     remove_unused_flags,
     set_auto_dipole,
     set_pmg_kpts,
 )
+from quacc.calculators.vasp.vasp_custodian import run_custodian
 from quacc.schemas.prep import set_magmoms
+from quacc.utils.dicts import sort_dict
 
 if TYPE_CHECKING:
     from typing import Literal
 
     from ase.atoms import Atoms
+    from pymatgen.io.vasp.sets import DictSet
 
 
 class Vasp(Vasp_):
@@ -50,6 +53,7 @@ class Vasp(Vasp_):
             | None
         ) = None,
         auto_dipole: bool | None = None,
+        pmg_input_set: DictSet | None = None,
         **kwargs,
     ) -> None:
         """
@@ -94,6 +98,9 @@ class Vasp(Vasp_):
         auto_dipole
             If True, will automatically set dipole moment correction parameters
             based on the center of mass (in the c dimension by default).
+        pmg_input_set
+            A Pymatgen input set to use for the VASP calculation, taken from a
+            `pymatgen.io.vasp.sets.DictSet` object.
         **kwargs
             Additional arguments to be passed to the VASP calculator, e.g.
             `xc='PBE'`, `encut=520`. Takes all valid ASE calculator arguments.
@@ -132,6 +139,7 @@ class Vasp(Vasp_):
         self.elemental_magmoms = elemental_magmoms
         self.pmg_kpts = pmg_kpts
         self.auto_dipole = auto_dipole
+        self.pmg_input_set = pmg_input_set
         self.kwargs = kwargs
 
         # Initialize for later
@@ -172,11 +180,6 @@ class Vasp(Vasp_):
                 "VASP_VDW setting was not provided, yet you requested a vdW functional."
             )
 
-        # Return Custodian executable command
-        if self.use_custodian:
-            run_vasp_custodian_file = Path(inspect.getfile(vasp_custodian)).resolve()
-            return f"{sys.executable} {run_vasp_custodian_file}"
-
         # Return vanilla ASE command
         vasp_cmd = (
             SETTINGS.VASP_GAMMA_CMD
@@ -208,6 +211,14 @@ class Vasp(Vasp_):
             msg = "Atoms object has a constraint that is not compatible with Custodian."
             raise ValueError(msg)
 
+        # Get Pymatgen VASP input set parameters
+        if self.pmg_input_set:
+            pmg_calc_params, self.input_atoms = get_pmg_input_set_params(
+                self.pmg_input_set, self.input_atoms
+            )
+        else:
+            pmg_calc_params = {}
+
         # Get user-defined preset parameters for the calculator
         if self.preset:
             calc_preset = load_vasp_yaml_calc(SETTINGS.VASP_PRESET_DIR / self.preset)[
@@ -218,7 +229,7 @@ class Vasp(Vasp_):
 
         # Collect all the calculator parameters and prioritize the kwargs in the
         # case of duplicates.
-        self.user_calc_params = calc_preset | self.kwargs
+        self.user_calc_params = pmg_calc_params | calc_preset | self.kwargs
 
         # Allow the user to use setups='mysetups.yaml' to load in a custom
         # setups from a YAML file
@@ -270,5 +281,43 @@ class Vasp(Vasp_):
             self.user_calc_params, self.pmg_kpts, self.input_atoms, self.incar_copilot
         )
 
-        # Remove unused INCAR flags
-        self.user_calc_params = remove_unused_flags(self.user_calc_params)
+        # Clean up the user calc parameters
+        self.user_calc_params = sort_dict(
+            normalize_params(remove_unused_flags(self.user_calc_params))
+        )
+
+    def _run(
+        self,
+        command: list[str] | None = None,
+        out: Path | str | None = None,
+        directory: Path | str = None,
+    ) -> int:
+        """
+        Override the Vasp calculator's run method to use Custodian if necessary.
+
+        Parameters
+        ----------
+        command
+            The command to run the VASP calculation. If None, will use the
+            self.command attribute.
+        out
+            The stdout file path.
+        directory
+            The directory to run the calculation in. If None, will use the
+            self.directory attribute.
+
+        Returns
+        -------
+        int
+            The return code.
+        """
+        if command is None:
+            command = self.command
+        if directory is None:
+            directory = self.directory
+
+        if self.use_custodian:
+            run_custodian()
+            return 0
+        else:
+            return subprocess.call(command, shell=True, stdout=out, cwd=directory)