+++ src/quacc/schemas/ase.py
@@ -96,7 +96,7 @@ def summarize_run(
     inputs = {
         "parameters": final_atoms.calc.parameters,
         "nid": uri.split(":")[0],
-        "dir_name": ":".join(uri.split(":")[1:]),
+        "dir_name": directory,
         "input_atoms": input_atoms_metadata,
         "quacc_version": __version__,
     }
@@ -286,7 +286,7 @@ def summarize_vib_run(
                 "nfree": vib.nfree,
             },
             "nid": uri.split(":")[0],
-            "dir_name": ":".join(uri.split(":")[1:]),
+            "dir_name": directory,
         }
 
     atoms_metadata = atoms_to_metadata(