+++ tests/core/utils/test_files.py
@@ -1,13 +1,19 @@
+import gzip
+import logging
 import os
 from pathlib import Path
 
 import pytest
 
 from quacc.utils.files import (
+    check_logfile,
     copy_decompress_files_from_dir,
     copy_decompress_tree,
     make_unique_dir,
 )
+
+LOGGER = logging.getLogger(__name__)
+LOGGER.propagate = True
 
 
 def test_make_unique_dir(tmp_path, monkeypatch):
@@ -97,3 +103,20 @@ def test_copy_decompress_files_from_dir_v2(tmp_path, monkeypatch):
     assert (dst / "dir1" / "file2").exists()
     assert Path(f"{dst}{'/nested' * 10}").exists()
     assert not (dst / "dir1" / "symlink1").exists()
+
+
+def test_copy_decompress_files_from_dir_v3(caplog):
+    with caplog.at_level(logging.WARNING):
+        copy_decompress_files_from_dir("fake", "test")
+
+
+def test_check_logfile(tmp_path):
+    with open(tmp_path / "logs.out", "w") as f:
+        f.write("trigger")
+    assert check_logfile(str(tmp_path / "logs.out"), "trigger") is True
+    assert check_logfile(str(tmp_path / "logs.out"), "test") is False
+
+    with gzip.open(tmp_path / "logs2.out.gz", "wb") as gf:
+        gf.write(b"trigger")
+    assert check_logfile(str(tmp_path / "logs2.out"), "trigger") is True
+    assert check_logfile(str(tmp_path / "logs2.out"), "test") is False